fix: store latency_sum as double, and stop scanning twice on retention
knex float maps to a 4 byte real on Postgres, which holds about seven
significant digits. latency_sum accumulates a sum, thus it is now a
double. max_latency and min_latency go with it, so the three latency
columns keep one type.

The retention path asked monitoring_data which tags it was about to
touch, then deleted, which meant two scans of the same set on the
largest table. It now walks the monitors table instead, which is small
and already indexed. Rebuilding a bucket for a monitor with no rows
behind it is a no-op, so the result is the same.

+++ migrations/20260906120000_add_monitoring_data_bucket.ts
@@ -26,11 +26,13 @@ export async function up(knex: Knex): Promise<void> {
     table.integer("count_of_down").notNullable().defaultTo(0);
     table.integer("count_of_degraded").notNullable().defaultTo(0);
     table.integer("count_of_maintenance").notNullable().defaultTo(0);
-    // Totals rather than an average, so buckets stay addable.
-    table.float("latency_sum").notNullable().defaultTo(0);
+    // Totals rather than an average, so buckets stay addable. Double, not float:
+    // knex float maps to a 4 byte real on Postgres, which holds about 7
+    // significant digits, and this column accumulates a sum.
+    table.double("latency_sum").notNullable().defaultTo(0);
     table.integer("latency_count").notNullable().defaultTo(0);
-    table.float("max_latency");
-    table.float("min_latency");
+    table.double("max_latency");
+    table.double("min_latency");
     table.primary(["monitor_tag", "bucket_ts"]);
     // Retention deletes by bucket_ts alone. The primary key leads with
     // monitor_tag, so it gives no useful path for that, and the nightly cleanup
